% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avanz.R
\name{avanz}
\alias{avanz}
\title{classe che rappresenta uno stato avanzamento}
\description{
classe che rappresenta uno stato avanzamento

classe che rappresenta uno stato avanzamento
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{avanz$new()}}
\item \href{#method-from_disk}{\code{avanz$from_disk()}}
\item \href{#method-to_disk}{\code{avanz$to_disk()}}
\item \href{#method-clone}{\code{avanz$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avanz$new(f)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_disk"></a>}}
\if{latex}{\out{\hypertarget{method-from_disk}{}}}
\subsection{Method \code{from_disk()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avanz$from_disk()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_disk"></a>}}
\if{latex}{\out{\hypertarget{method-to_disk}{}}}
\subsection{Method \code{to_disk()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avanz$to_disk()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avanz$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
