% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avanz.R
\name{avanz}
\alias{avanz}
\title{classe che rappresenta uno stato avanzamento}
\description{
classe che rappresenta uno stato avanzamento

classe che rappresenta uno stato avanzamento
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{avanz$new()}}
\item \href{#method-from_disk}{\code{avanz$from_disk()}}
\item \href{#method-to_disk}{\code{avanz$to_disk()}}
\item \href{#method-setup}{\code{avanz$setup()}}
\item \href{#method-assignable_files}{\code{avanz$assignable_files()}}
\item \href{#method-list_assignee}{\code{avanz$list_assignee()}}
\item \href{#method-unfinished_homeworks}{\code{avanz$unfinished_homeworks()}}
\item \href{#method-assign}{\code{avanz$assign()}}
\item \href{#method-clone}{\code{avanz$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avanz$new(f)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_disk"></a>}}
\if{latex}{\out{\hypertarget{method-from_disk}{}}}
\subsection{Method \code{from_disk()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avanz$from_disk()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_disk"></a>}}
\if{latex}{\out{\hypertarget{method-to_disk}{}}}
\subsection{Method \code{to_disk()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avanz$to_disk(newdata = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setup"></a>}}
\if{latex}{\out{\hypertarget{method-setup}{}}}
\subsection{Method \code{setup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avanz$setup(trn_filenames)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assignable_files"></a>}}
\if{latex}{\out{\hypertarget{method-assignable_files}{}}}
\subsection{Method \code{assignable_files()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avanz$assignable_files(role = c("translator", "revisor1", "revisor2"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_assignee"></a>}}
\if{latex}{\out{\hypertarget{method-list_assignee}{}}}
\subsection{Method \code{list_assignee()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avanz$list_assignee()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unfinished_homeworks"></a>}}
\if{latex}{\out{\hypertarget{method-unfinished_homeworks}{}}}
\subsection{Method \code{unfinished_homeworks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avanz$unfinished_homeworks(u, role = c("translator", "revisor1", "revisor2"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assign"></a>}}
\if{latex}{\out{\hypertarget{method-assign}{}}}
\subsection{Method \code{assign()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avanz$assign(
  old_f,
  assignee,
  new_f,
  role = c("translator", "revisor1", "revisor2")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avanz$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
