% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj.R
\name{prj}
\alias{prj}
\title{classe che rappresenta un progetto}
\description{
classe che rappresenta un progetto

classe che rappresenta un progetto
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{prj$new()}}
\item \href{#method-setup}{\code{prj$setup()}}
\item \href{#method-create_sandbox}{\code{prj$create_sandbox()}}
\item \href{#method-assign}{\code{prj$assign()}}
\item \href{#method-mark_progresses}{\code{prj$mark_progresses()}}
\item \href{#method-make_final_srt}{\code{prj$make_final_srt()}}
\item \href{#method-list_assignee}{\code{prj$list_assignee()}}
\item \href{#method-monitoring}{\code{prj$monitoring()}}
\item \href{#method-clone}{\code{prj$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prj$new(id = NULL, yt_id = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setup"></a>}}
\if{latex}{\out{\hypertarget{method-setup}{}}}
\subsection{Method \code{setup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prj$setup(chunks_len_mins = 5, trn_to_rev_ratio = 6)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_sandbox"></a>}}
\if{latex}{\out{\hypertarget{method-create_sandbox}{}}}
\subsection{Method \code{create_sandbox()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prj$create_sandbox(sandbox_f = "", rev1_sandbox_f = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assign"></a>}}
\if{latex}{\out{\hypertarget{method-assign}{}}}
\subsection{Method \code{assign()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prj$assign(translate_f = "", revise2_f = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mark_progresses"></a>}}
\if{latex}{\out{\hypertarget{method-mark_progresses}{}}}
\subsection{Method \code{mark_progresses()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prj$mark_progresses(
  trn_completed_f = "",
  rev1_started_f = "",
  rev1_completed_f = "",
  rev2_completed_f = ""
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_final_srt"></a>}}
\if{latex}{\out{\hypertarget{method-make_final_srt}{}}}
\subsection{Method \code{make_final_srt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prj$make_final_srt()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_assignee"></a>}}
\if{latex}{\out{\hypertarget{method-list_assignee}{}}}
\subsection{Method \code{list_assignee()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prj$list_assignee()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-monitoring"></a>}}
\if{latex}{\out{\hypertarget{method-monitoring}{}}}
\subsection{Method \code{monitoring()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prj$monitoring()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prj$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
